+++ sample/test/ServerIntegrationTest.kt
@@ -14,10 +14,9 @@ import java.net.http.HttpRequest.BodyPublishers.noBody
 
 class ServerIntegrationTest {
   @Test fun requests() {
-    val port = (Math.random() * 60000 + 1000).toInt()
-    val server = sampleServer(port).apply { start(gracefulStopDelaySec = -1) }
+    val server = sampleServer(0).apply { start(gracefulStopDelaySec = -1) }
 
-    val http = JsonHttpClient("http://localhost:$port", registry = server.registry)
+    val http = JsonHttpClient("http://localhost:${server.listen.port}", registry = server.registry)
     runBlocking {
       expect(http.get<String>("/hello")).toEqual("\"Hello World\"")
       expect(http.get<String>("/hello/param/123456?query=123")).toEqual("\"Path: 123456, Query: {query=123}\"")